Update tests and snapshots.
